# contracts/openapi-v1.json
{
  "components": {
    "schemas": {
      "ClientSettings": {
        "additionalProperties": false,
        "properties": {
          "auto_suggest": {
            "default": true,
            "title": "Auto Suggest",
            "type": "boolean"
          },
          "diagnostics": {
            "default": false,
            "title": "Diagnostics",
            "type": "boolean"
          },
          "save_history": {
            "default": true,
            "title": "Save History",
            "type": "boolean"
          },
          "theme": {
            "default": "cream",
            "enum": [
              "cream",
              "soda",
              "moon"
            ],
            "title": "Theme",
            "type": "string"
          }
        },
        "title": "ClientSettings",
        "type": "object"
      },
      "ConversationTurn": {
        "additionalProperties": false,
        "properties": {
          "role": {
            "enum": [
              "self",
              "other"
            ],
            "title": "Role",
            "type": "string"
          },
          "text": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "role",
          "text"
        ],
        "title": "ConversationTurn",
        "type": "object"
      },
      "GuestMergeRequest": {
        "additionalProperties": false,
        "properties": {
          "guest_refresh_token": {
            "maxLength": 4096,
            "minLength": 40,
            "title": "Guest Refresh Token",
            "type": "string"
          }
        },
        "required": [
          "guest_refresh_token"
        ],
        "title": "GuestMergeRequest",
        "type": "object"
      },
      "GuestSessionRequest": {
        "additionalProperties": false,
        "properties": {
          "app_version": {
            "pattern": "^(0|[1-9]\\d*)\\.(0|[1-9]\\d*)\\.(0|[1-9]\\d*)(?:-[0-9A-Za-z.-]+)?$",
            "title": "App Version",
            "type": "string"
          },
          "device_id": {
            "maxLength": 64,
            "minLength": 8,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Device Id",
            "type": "string"
          },
          "platform": {
            "enum": [
              "android",
              "windows"
            ],
            "title": "Platform",
            "type": "string"
          }
        },
        "required": [
          "device_id",
          "platform",
          "app_version"
        ],
        "title": "GuestSessionRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "MergeResponse": {
        "additionalProperties": false,
        "properties": {
          "status": {
            "const": "merged",
            "default": "merged",
            "title": "Status",
            "type": "string"
          }
        },
        "title": "MergeResponse",
        "type": "object"
      },
      "RefreshTokenRequest": {
        "additionalProperties": false,
        "properties": {
          "refresh_token": {
            "maxLength": 4096,
            "minLength": 40,
            "title": "Refresh Token",
            "type": "string"
          }
        },
        "required": [
          "refresh_token"
        ],
        "title": "RefreshTokenRequest",
        "type": "object"
      },
      "ReleaseManifest": {
        "additionalProperties": false,
        "properties": {
          "mandatory": {
            "title": "Mandatory",
            "type": "boolean"
          },
          "minimum_supported_version": {
            "pattern": "^(0|[1-9]\\d*)\\.(0|[1-9]\\d*)\\.(0|[1-9]\\d*)(?:-[0-9A-Za-z.-]+)?$",
            "title": "Minimum Supported Version",
            "type": "string"
          },
          "platform": {
            "enum": [
              "android",
              "windows"
            ],
            "title": "Platform",
            "type": "string"
          },
          "published_at": {
            "format": "date-time",
            "title": "Published At",
            "type": "string"
          },
          "sha256": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Sha256",
            "type": "string"
          },
          "signature": {
            "maxLength": 4096,
            "minLength": 8,
            "title": "Signature",
            "type": "string"
          },
          "size": {
            "exclusiveMinimum": 0.0,
            "title": "Size",
            "type": "integer"
          },
          "url": {
            "format": "uri",
            "minLength": 1,
            "title": "Url",
            "type": "string"
          },
          "version": {
            "pattern": "^(0|[1-9]\\d*)\\.(0|[1-9]\\d*)\\.(0|[1-9]\\d*)(?:-[0-9A-Za-z.-]+)?$",
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "platform",
          "version",
          "minimum_supported_version",
          "mandatory",
          "url",
          "size",
          "sha256",
          "signature",
          "published_at"
        ],
        "title": "ReleaseManifest",
        "type": "object"
      },
      "SmsSendRequest": {
        "additionalProperties": false,
        "properties": {
          "phone": {
            "pattern": "^\\+[1-9]\\d{7,14}$",
            "title": "Phone",
            "type": "string"
          }
        },
        "required": [
          "phone"
        ],
        "title": "SmsSendRequest",
        "type": "object"
      },
      "SmsSendResponse": {
        "additionalProperties": false,
        "properties": {
          "retry_after": {
            "default": 60,
            "title": "Retry After",
            "type": "integer"
          },
          "status": {
            "const": "sent",
            "default": "sent",
            "title": "Status",
            "type": "string"
          }
        },
        "title": "SmsSendResponse",
        "type": "object"
      },
      "SmsVerifyRequest": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "pattern": "^\\d{6}$",
            "title": "Code",
            "type": "string"
          },
          "device_id": {
            "maxLength": 64,
            "minLength": 8,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Device Id",
            "type": "string"
          },
          "phone": {
            "pattern": "^\\+[1-9]\\d{7,14}$",
            "title": "Phone",
            "type": "string"
          }
        },
        "required": [
          "phone",
          "code",
          "device_id"
        ],
        "title": "SmsVerifyRequest",
        "type": "object"
      },
      "SuggestionRequest": {
        "additionalProperties": false,
        "properties": {
          "application": {
            "enum": [
              "wechat",
              "qq",
              "wecom",
              "dingtalk",
              "feishu",
              "whatsapp",
              "telegram",
              "discord",
              "generic"
            ],
            "title": "Application",
            "type": "string"
          },
          "device_id": {
            "maxLength": 64,
            "minLength": 8,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Device Id",
            "type": "string"
          },
          "draft": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Draft"
          },
          "locale": {
            "maxLength": 16,
            "minLength": 2,
            "title": "Locale",
            "type": "string"
          },
          "request_id": {
            "maxLength": 64,
            "minLength": 8,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Request Id",
            "type": "string"
          },
          "save_history": {
            "default": true,
            "title": "Save History",
            "type": "boolean"
          },
          "turns": {
            "items": {
              "$ref": "#/components/schemas/ConversationTurn"
            },
            "maxItems": 12,
            "minItems": 1,
            "title": "Turns",
            "type": "array"
          }
        },
        "required": [
          "request_id",
          "device_id",
          "application",
          "locale",
          "turns"
        ],
        "title": "SuggestionRequest",
        "type": "object"
      },
      "TokenPair": {
        "additionalProperties": false,
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "refresh_token": {
            "title": "Refresh Token",
            "type": "string"
          },
          "subject_type": {
            "enum": [
              "guest",
              "user"
            ],
            "title": "Subject Type",
            "type": "string"
          },
          "token_type": {
            "const": "bearer",
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "refresh_token",
          "expires_in",
          "subject_type"
        ],
        "title": "TokenPair",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Langou API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/ai/suggestions:stream": {
      "post": {
        "operationId": "stream_suggestions_v1_ai_suggestions_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuggestionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Stream Suggestions"
      }
    },
    "/v1/auth/sms/merge": {
      "post": {
        "operationId": "merge_guest_v1_auth_sms_merge_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuestMergeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MergeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Merge Guest"
      }
    },
    "/v1/auth/sms/send": {
      "post": {
        "operationId": "send_sms_v1_auth_sms_send_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SmsSendRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SmsSendResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Send Sms"
      }
    },
    "/v1/auth/sms/verify": {
      "post": {
        "operationId": "verify_sms_v1_auth_sms_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SmsVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Sms"
      }
    },
    "/v1/auth/token/refresh": {
      "post": {
        "operationId": "refresh_token_v1_auth_token_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh Token"
      }
    },
    "/v1/devices/guest-session": {
      "post": {
        "operationId": "create_guest_session_v1_devices_guest_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuestSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Guest Session"
      }
    },
    "/v1/history": {
      "delete": {
        "operationId": "delete_history_v1_history_delete",
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "minLength": 8,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete History"
      },
      "get": {
        "operationId": "list_history_v1_history_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List History V1 History Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List History"
      }
    },
    "/v1/releases/{platform}/latest": {
      "get": {
        "operationId": "latest_release_v1_releases__platform__latest_get",
        "parameters": [
          {
            "in": "path",
            "name": "platform",
            "required": true,
            "schema": {
              "enum": [
                "android",
                "windows"
              ],
              "title": "Platform",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseManifest"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Latest Release"
      }
    },
    "/v1/settings": {
      "get": {
        "operationId": "get_settings_v1_settings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientSettings"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Settings"
      },
      "put": {
        "operationId": "put_settings_v1_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientSettings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientSettings"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Put Settings"
      }
    }
  }
}